requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
coloredlogs==15.0.1
frozenlist==1.4.1
humanfriendly==10.0
idna==3.7
multidict==6.0.5
pillow==10.4.0
pyodbc==5.1.0
requests==2.32.3
urllib3==2.2.2
yarl==1.9.4
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
coloredlogs==15.0.1
frozenlist==1.4.1
humanfriendly==10.0
idna==3.7
multidict==6.0.5
pillow==10.4.0
pyodbc==5.1.0
requests==2.32.3
urllib3==2.2.2
yarl==1.9.4
